app/terminal: merge `waitstatus.js` into `terminal.js`
